+++ src/felupe/thermal/_solidbody_thermal.py
@@ -136,7 +136,7 @@ class SolidBodyThermal(SolidBody):
     --------
     felupe.thermal.TimeStep : A time step item.
     felupe.thermal.SolidBodySurfaceHeatTransfer : A thermal convection boundary condition.
-    felupe.thermal.SolidBodyThermalHeatFlux : A thermal heat flux boundary condition.
+    felupe.thermal.SolidBodyHeatFlux : A thermal heat flux boundary condition.
 
     """